[extimg] extimg and silx-extimg share IO routines
